// src/retentions/dto/retention.dto.ts
export class RetentionDto {
  code: string;
  descripction: string;
  reg_hold_ind: string;
  trans_hold_ind: string;
  grad_hold_ind: string;
  grade_hold_ind: string;
  ar_hold_ind: string;
  env_hold_ind: string;
  vr_msg_no: string;
  disp_web_ind: string;
  application_hold_ind: string;
  compliance_hold_ind: string;
  user_id: string;
  data_origin: string;
  activity_date: string;
}